début d'implémentation des quadrants des statistiques dans la page Analyse
